requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
billiard==4.2.2
black==25.1.0
celery==5.5.3
certifi==2025.11.12
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cache2==0.2.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
iniconfig==2.3.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pillow==12.0.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyright==1.1.407
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==7.0.1
ruff==0.14.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.14
